rename and refactor according to review tips
